Fix invalid number usage in oauth variable
Recent changes to the oauth token maxage variable introduced it as a
number instead of a string or integer, which SCAP is expecting. This
commit updates the variable to be a string instead of a number type, so
that the resulting datastream is valid and parsable by the compliance
oeprator profile parsers.

Without this change, content from the latest branch will cause the
compliance operator installs to hang because it can't parse the xccdf.