OCP4: use utf-8 as default xml encoding
This commit make UTF-8 as default xml encoding. In a recent change, when writing xccdf xml file, it uses the output encoding US-ASCII as default if not set, however this causes issues in Compliance Operator parser, using US-ASCII, the parser is not able to get the prefix of xml node, therefore crash the parser pod.