change the OVAL for sssd-ldap platform
Originally the platform was ipplicable only in case sssd was configured to use active directory.
There are two cases when sssd can use LDAP.
When ldap is specified explicitly or when using ipa.
However, it is not recommended to add manual settings to the sssd configuration when ipa is used.