Fix handling of grub.d configs in grub2_bootloader_argument
Fixed failing check on Ubuntu when grub parameters were defined in
/etc/default/grub.d/*cfg but not in /etc/default/grub.

The fixed oval checks for the existence of the parameter
in *either* /etc/default/grub or *at least one*
/etc/default/grub.d/*cfg, ignoring loading order or overriden values.

We can avoid the added complexity of checking the loading
order and potentially overridden GRUB_CMDLINE_LINUX variables,
because the tests with /etc/default/grub* are redundant to begin with.
The OVAL checks also the parameters in the generated grub configs
in /boot (as prescribed by CIS 1.0.0 for Ubuntu 22.04).