OCPBUGS-13589: Remove sysctl file check rule from STIG profile
We already have a bunch of rules that check the values of kernel sysctl
settings alongside a rule the ensures the default sysctl settings are
used. But we also had a rule to ensure the sysctl kernel parameter
configuration file was on disk for each node in the cluster.

The rules are structured so that we only need to check if the defaults
are being used or that all individual checks are being satified. By
default the sysctl rules pass on OpenShift 4.13+.

This makes the sysctl file existence check obsolete, and confusing for
users because the file doesn't exist by default. This causes the rule to
fail because the file doesn't exist. Remediations for other sysctl rules
will create the file, but since they pass by default the remediations
are not generated. To get around this, the user either need to exclude
the rule in a tailored profile, or they need to put a specific file on
disk that doesn't do anything.

This commit removes the rule from the STIG, and relies on checking the the
default kernel parameters. This commit also removes sub-rules from the
profile, since those rules are essentially helper rules for
`kubelet_enable_protect_kernel_sysctl`, and we don't need to specify
them twice. So even though we're removing some rules from the profile,
we're not decreasing coverage.