Enforce explicit setting in system-auth
The rule passes if the `rounds` option isn't set but the profile
uses a value that equals to system default value. This will no
longer be possible. We will always require `rounds` to be set
explicitly. The reason is to align the rule with DISA STIG.

The commit also updates the test scenarios to fail if the
`rounds` isn't set and system default is used.

Fixes: #11697