Do not require existence of /var/tmp/tmp-inst
We will not require the parent directory to exist, the OVAL test will
pass also if the directory doesn't exist.  It isn't mandatory to create
the parent directory because when the directory doesn't exist, it gets
created automatically by pam. However, if the parent directory exists,
it must have correct mode, otherwise the polyinstantiation will fail.

This change is very similar to the change that we have done to the
sister rule accounts_polyinstantiated_tmp in 13d2d86.

More context:
https://github.com/RHSecurityCompliance/contest/pull/129#issuecomment-2019515006