Fix Ansible in chronyd_or_ntpd_set_maxpoll
The Ansible Tasks in this rule manipulate with many unrelated files such
as /etc/sestatus.conf or /etc/krb5.conf.  The reason is that they
wrongly use the chrony_conf_path variable. They get the name of the
parent directory and look for all .conf files there. This probably works
nicely for Ubuntu products, where chrony_conf_path is set to
/etc/chrony/chrony.conf, but creates harm in all other products where
chrony_conf_path is set to /etc/chrony.conf, so the dirname is /etc and
the search matches all *.conf files in /etc.

We will fix this problem by explicit using the chrony_d_path content
variable which defines path to the chrony configuration directory.