Disable remediation for smartcard_pam_enabled on Ubuntu 22.04
The current remediation is not considered reliable enough for
critical PAM configuration and will be disabled until a better
solution is implemented. The remediation breaks the PAM auth
configuration on a clean Ubuntu Desktop 22.04 installation which ships
with `libpam-sss` (and thus `auth [success=2...] libpam_sss.so`)
by default.

The PAM rules should ideally rely on `pam-auth-update` as was
implemented in recent CIS releases.