Add subscription file for ROSA installs
Since ROSA clusters only have worker nodes, we need to modify the
subscription to allow the operator to schedule on worker nodes.

Without this, ROSA installations will spin waiting for master nodes to
schedule on.

This will get wired up to CI through the ocp4e2e test suite in a
separate PR:

  https://github.com/ComplianceAsCode/ocp4e2e/pull/41